moving launcher code from treasure.tsx to useLauncher.tsx
